requirements.txt
# fastapi_backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary # PostgreSQL 드라이버
sqlalchemy # ORM
alembic # 데이터베이스 마이그레이션 도구
pydantic[email] # 데이터 유효성 검사 (이메일 유효성 포함)
pydantic-settings # Pydantic v2+ 환경 변수 로딩용
python-jose[cryptography] # JWT 처리
passlib[bcrypt] # 비밀번호 해싱
python-multipart # 파일 업로드 처리
firebase-admin # Firebase Admin SDK (Firebase ID 토큰 검증용)